Move config.json/meta.json to user_files and convert to turtle format
There's no reason for the plugins internal representation of the notes to be configurable. In fact, if a user were to edit meta.json (as they could be expected to) then the plugin would break some or all of the integrations with the user's notes.

By moving over to user_files, we're much more in control of how we want to store the data, instead of being arbitrarily being restricted to a single JSON file.