Remove sys.path.append in imports
This was originally added such that symlinks work so you can use your normal Anki installation to run the program.
I've found this to be a hassle more than anything, as the installed Anki installation needs to match the assumed version of this add-on.
It causes obscure bugs which aren't worth the hassle